requirements.txt
Pillow
scipy
matplotlib
numpy
xmltodict